+++ cello.cpp
@@ -27,6 +27,7 @@
 #error "Incorrect use of JUCE cpp file"
 #endif
 
+#include "cello/cello_computed_value.cpp"
 #include "cello/cello_ipc.cpp"
 #include "cello/cello_object.cpp"
 #include "cello/cello_path.cpp"