+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased 
+
+### Added
+### Changed
+### Fixed
+
+## 1.7.0 * 2025-12-17
+
+### Added 
+
+- `Object::remove(const cello::Query&)` will remove any children that are filtered by the query argument, returning the number of children that were removed as a result. 
+- new `ComputedValue<T>` class; provides a `Value`-like API to functions/data that are not backed in the ValueTree as such (useful for conversions, etc.)
+
+### Changed
+
+### Fixed 
+
 ## 1.6.0 * 2025-08-21
 
 ### Added